Web platform revision

+++ iDMED/src/components/Patient/PatientPanel/AddEditPrescription.vue
@@ -348,11 +348,13 @@ export default {
     async init () {
       this.setStep(this.stepp)
       this.clearPrescriptionSession()
-      ClinicalServiceAttribute.deleteAll()
-      await ClinicalServiceAttribute.localDbGetAll().then(regimens => {
-        console.log(regimens)
-        ClinicalServiceAttribute.insert({ data: regimens })
-      })
+      if (this.mobile) {
+        ClinicalServiceAttribute.deleteAll()
+        await ClinicalServiceAttribute.localDbGetAll().then(regimens => {
+          console.log(regimens)
+          ClinicalServiceAttribute.insert({ data: regimens })
+        })
+      }
       if (this.isNewPackStep || this.isEditPackStep) {
         this.initPatientVisitDetailsForDispense()
       } else {
@@ -921,8 +923,11 @@ export default {
           }
         }.bind(this))
 
-        const i = 0
-        this.saveVisitPrescriptionAndPack(this.patientVisit, i)
+        // const i = 0
+        // this.saveVisitPrescriptionAndPack(this.patientVisit, i)
+        this.patientVisit.patient = this.simplePatient
+        this.patientVisit.clinic = this.currClinic
+        PatientVisit.apiSave(JSON.parse(JSON.stringify(this.patientVisit)))
       }
     },
     saveVisitPrescriptionAndPack (patientVisit, i) {
